functions/upload-image/requirements.txt
google-cloud-storage
google-cloud-pubsub
flask

